export group function
