+++ src/types/bancor.d.ts
@@ -272,6 +272,36 @@ export interface ViewReserve {
   reserveWeight: number;
 }
 
+export interface ViewGroupedPositions {
+  id: string;
+  poolId: string;
+  symbol: string;
+  stake: {
+    amount: number;
+    usdValue: number;
+    unixTime: number;
+  };
+  protectedAmount: {
+    amount: number;
+    usdValue: number;
+  };
+  fullyProtected: {
+    amount: number;
+    usdValue: number;
+  };
+  apr: {
+    day: number;
+    week: number;
+    // month: number;
+  };
+  fees: number;
+  roi: number;
+  insuranceStart: number;
+  coverageDecPercent: number;
+  fullCoverage: number;
+  collapsedData: ViewProtectedLiquidity[];
+}
+
 export interface ViewRelay {
   id: string;
   symbol: string;
